Revert "Fix:: 컴파일 에러 수정"

See merge request s10-webmobile1-sub2/S10P12E107!37

+++ back-gollajyu/src/main/java/com/jaecheop/backgollajyu/comment/model/CommentResDto.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class CommentResDto {
 
-    private Long commentId;
+    private int commentId;
     private LocalDateTime commentCreateAt;
     private String commentDescription;
     private VoteResDto voteResDto;
